Remove upstreamed patches, python helpers and scripts
Patch17 in 9b4c50dd67c337f2d3c927cdd01ae4433bb08b61
Patch20 in c6ff61448a4ead0540130c9a6eb052fcd71b2624
Patch22 moved to python-srpm-macros
Patch23 in e0de325e1d13fb65066ad09d16cb08ba6e294901
Patch26 in 04b0805a756cdd9466fb9dc717846f5bf15518cc
Patch27 in 04b0805a756cdd9466fb9dc717846f5bf15518cc
Python scripts in bc4829771ca4e0ce4e1d9d8d91afea4f30c3ea53

https://github.com/rpm-software-management/python-rpm-packaging/

+++ rpm/rpm.spec
@@ -127,6 +127,12 @@ export CPPFLAGS CFLAGS LDFLAGS
 
 find %{buildroot} -regex ".*\\.la$" | xargs rm -f --
 
+# Database backend is auto-detected during build
+if ! grep -E '^%%_db_backend[[:space:]]+ndb$' ${RPM_BUILD_ROOT}%{_rpmconfigdir}/macros; then
+    echo "Default database is not ndb"
+    exit 1
+fi
+
 # We cannot use _unitdir macro as we don't want to depend on systemd
 mkdir -p $RPM_BUILD_ROOT/usr/lib/systemd/system
 install -m 644 %{SOURCE2} $RPM_BUILD_ROOT/usr/lib/systemd/system
